+++ IOS-10 HW-13-Kalyk Azhar/Sources/Module/View.swift
@@ -1,0 +1,49 @@
+import UIKit
+import SnapKit
+
+final class MainView: UIView {
+    
+    // MARK: - Elements
+    
+    lazy var tableView: UITableView = {
+        let tableView = UITableView(frame: .zero, style: .insetGrouped)
+        tableView.register(CustomTableViewCell.self, forCellReuseIdentifier: "cell")
+        
+        return tableView
+    }()
+    
+    
+    
+    // MARK: - Lifecycle
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setupView()
+        setupHierarchy()
+        setupLayout()
+        
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError()
+    }
+    
+    // MARK: - Setup
+    
+    private func setupView() {
+        backgroundColor = .white
+    }
+    
+    private func setupHierarchy() {
+        addSubview(tableView)
+    }
+    
+    private func setupLayout() {
+        
+        tableView.snp.makeConstraints { make in
+            make.top.equalTo(safeAreaLayoutGuide).offset(20)
+            make.right.bottom.left.equalToSuperview()
+        }
+        
+    }
+}
